рефактор стилей ком-ов Header, Main, PageWrapper (изменил пути картинок services/mock-post-controller)
