+++ src/common/pageWrapper/PageWrapper.jsx
@@ -1,10 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import styled from 'styled-components';
 import Header from '../header';
 import LeftBlock from '../leftBlock';
 import RightBlock from '../rightBlock';
-import { MainContainer } from '../styledComponents';
 
+const MainContainer = styled.div`
+  display: flex;
+  width: 100%;
+  background: #111;
+`;
 const PageWrapper = ({ messages, children }) => (
   <>
     <Header />
