+++ src/common/pageWrapper/index.js
@@ -1,0 +1,3 @@
+import PageWrapper from './PageWrapper';
+
+export default PageWrapper;
