+++ tsconfig.json
@@ -13,7 +13,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react",
+    "jsx": "react-jsx",
     "baseUrl": "src",
     "typeRoots": ["./typings", "./node_modules/@types"],
     "noFallthroughCasesInSwitch": true
